+++ src/challenges/chess-simp.ts
@@ -1,7 +1,7 @@
 import { Uuid } from '@/utils/uuid'
 import _ from 'lodash'
 import { Simp_2021_08_17 } from './chess-simp/2021-08'
-import { Simp_2021_10_16 } from './chess-simp/2021-10'
+import { Simp_2021_10_16, Simp_2021_10_22 } from './chess-simp/2021-10'
 import { Simp_2021_12_04 } from './chess-simp/2021-12'
 import { Simp_2022_01_21, Simp_2022_01_29 } from './chess-simp/2022-01'
 import { Simp_2022_02_10, Simp_2022_02_11 } from './chess-simp/2022-02'
@@ -36,7 +36,7 @@ export const chessSimpChallenges: Map<Uuid, { meta: ChallengeMeta; create: () =>
   new Map(
     _.concat(
       [() => new Simp_2021_08_17() as Challenge],
-      [() => new Simp_2021_10_16() as Challenge],
+      [() => new Simp_2021_10_16() as Challenge, () => new Simp_2021_10_22() as Challenge],
       [() => new Simp_2021_12_04() as Challenge],
       [() => new Simp_2022_01_29() as Challenge, () => new Simp_2022_01_21() as Challenge],
       [() => new Simp_2022_02_10() as Challenge, () => new Simp_2022_02_11() as Challenge],